backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
SQLAlchemy==2.0.32
PyMySQL==1.1.1
python-dotenv==1.0.1
gunicorn
